backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
anthropic==0.40.0
httpx[socks]==0.27.2
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
aiosqlite==0.20.0

